begin adding roofing, starting with the three variations of tin
