+++ jv_utils.py
@@ -3,6 +3,6 @@ class Units:
     INCH: float = FOOT / 12
     TQ_INCH: float = INCH * 0.75  # 3/4 inch
     H_INCH: float = INCH / 2  # 1/2 inch
-    Q_INCH: float = H_INCH / 2 # 1/4 inch
+    Q_INCH: float = H_INCH / 2  # 1/4 inch
     ETH_INCH: float = INCH / 8  # 1/8th inch
     STH_INCH: float = INCH / 16  # 1/16th inch
